Backend/requirements.txt
# Core
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
sqlalchemy==1.4.46
python-jose==3.3.0
passlib==1.7.4
greenlet==2.0.2
typing-extensions==4.12.0
h11==0.14.0  

alembic==1.12.1

# Machine Learning y Data Science
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2

# Validación de datos
pydantic==2.5.0

# Autenticación y seguridad
python-dotenv==1.0.0


# Testing (opcional)
pytest==7.4.3
httpx==0.25.2

# Logging y monitoreo
loguru==0.7.2

# Utilidades
python-dateutil==2.8.2
